fixed a printf problem
git-svn-id: https://groke.otago.ac.nz/ant/ant@1071 7489a7b4-4588-4892-8acd-774bacfab51c

+++ source/query_parse_tree.c
@@ -30,7 +30,7 @@ printf("(");
 switch (boolean_operator)
 	{
 	case LEAF_NODE:
-		printf("%*.*s", term.string_length, term.string_length, term.start);
+		printf("%*.*s", (int)term.string_length, (int)term.string_length, term.start);
 		break;
 	case BOOLEAN_AND:
 		left->internal_text_render();
